# py/hmm_service/docs/api_specification.yaml
openapi: 3.0.3
info:
  title: HMM Microservice API
  description: |
    Hidden Markov Model inference service for the IMP trading system.
    
    This service provides real-time HMM state probability calculation and fusion weight computation
    for regime-aware signal fusion in trading applications.
    
    ## Features
    - State probability calculation from observation vectors
    - Fusion weight computation for signal combination
    - Model hot-reloading without service downtime
    - Comprehensive health monitoring and metrics
    - Structured error handling with detailed diagnostics
    
    ## Performance Targets
    - Inference latency: <20ms (p95)
    - Health check latency: <5ms
    - Concurrent requests: 100+ req/sec
    - Availability: 99.9%
    
    ## Authentication
    Currently no authentication is required. In production, consider implementing:
    - API key authentication
    - IP whitelisting
    - Rate limiting per client
  version: 1.0.0
  contact:
    name: IMP Development Team
    email: dev@imp-trading.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://hmm-service.imp-trading.com
    description: Production server
paths:
...
  /models/reload:
    post:
      tags:
        - Models
      summary: Reload HMM model
      description: |
        Hot-reload HMM model from MinIO storage without service downtime.
        
        Supports:
        - Loading specific experiment IDs and versions
        - Model validation before activation
        - Fallback to previous model on failure
        - Atomic model switching
        
        The service continues operating with the previous model if reload fails.
      operationId: reloadModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelReloadRequest'
            examples:
              reload_latest:
                summary: Reload latest model version
                value:
                  experiment_id: "production_hmm"
                  version: "latest"
                  validate_model: true
              reload_specific:
                summary: Reload specific model version
                value:
                  experiment_id: "hmm_v2"
                  version: "1.2.0"
                  validate_model: true
      responses:
        '200':
          description: Model reloaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelReloadResponse'
              examples:
                successful_reload:
                  summary: Successful model reload
                  value:
                    success: true
                    model_info:
                      experiment_id: "production_hmm"
                      version: "1.2.0"
                      load_time: 1640995200
                      n_states: 3
                      n_features: 3
                      has_fusion_weights: true
                      validation_passed: true
                    reload_time: 2.5
                    previous_model:
                      experiment_id: "production_hmm"
                      version: "1.1.0"
                      load_time: 1640991600
                    timestamp: 1640995200
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/ModelNotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /models/current:
    get:
      tags:
        - Models
      summary: Get current model info
      description: |
        Get detailed information about the currently loaded HMM model.
        
        Includes:
        - Model metadata (states, features, version)
        - Performance statistics
        - Training information
        - Load timestamp
        
        Returns 404 if no model is currently loaded.
      operationId: getCurrentModel
      responses:
        '200':
          description: Current model information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentModelResponse'
              examples:
                current_model:
                  summary: Current model information
                  value:
                    experiment_id: "production_hmm"
                    version: "1.2.0"
                    load_time: 1640995200
                    model_info:
                      n_states: 3
                      n_features: 3
                      has_fusion_weights: true
                      library: "hmmlearn"
                      training_window:
                        start_date: "2024-01-01"
                        end_date: "2024-01-31"
                      training_samples: 50000
                      validation_score: 0.85
                      artifact_size_mb: 2.5
                    performance_stats:
                      total_inferences: 12500
                      avg_inference_time_ms: 15.2
                      error_rate: 0.001
                      cache_hit_rate: 0.75
                    timestamp: 1640995200
        '404':
          $ref: '#/components/responses/ModelNotFound'
  /models/available:
    get:
      tags:
        - Models
      summary: List available models
      description: |
        List all available HMM models in MinIO storage.
        
        Returns metadata for all models including:
        - Experiment IDs and versions
        - Creation timestamps
        - Model sizes and parameters
        - Validation scores
        
        Useful for model management and deployment planning.
      operationId: listAvailableModels
      responses:
        '200':
          description: Available models list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailableModelsResponse'
              examples:
                available_models:
                  summary: List of available models
                  value:
                    models:
                      - experiment_id: "production_hmm"
                        version: "1.2.0"
                        created_at: 1640995200
                        size_mb: 2.5
                        n_states: 3
                        n_features: 3
                        validation_score: 0.85
                        has_fusion_weights: true
                        library: "hmmlearn"
                      - experiment_id: "production_hmm"
                        version: "1.1.0"
                        created_at: 1640991600
                        size_mb: 2.3
                        n_states: 3
                        n_features: 3
                        validation_score: 0.82
                        has_fusion_weights: true
                        library: "hmmlearn"
                    total_count: 2
                    timestamp: 1640995200
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  schemas:
...
    ModelReloadRequest:
      type: object
      properties:
        experiment_id:
          type: string
          description: "Experiment ID to load (uses default if not specified)"
          example: "production_hmm"
        version:
          type: string
          description: "Model version to load"
          default: "latest"
          example: "1.2.0"
        validate_model:
          type: boolean
          description: "Whether to validate model before activation"
          default: true
          example: true
    ModelReloadResponse:
      type: object
      required:
        - success
        - model_info
        - reload_time
        - timestamp
      properties:
        success:
          type: boolean
          description: "Whether reload was successful"
          example: true
        model_info:
          type: object
          description: "Information about loaded model"
        reload_time:
          type: number
          format: float
          minimum: 0.0
          description: "Time taken to reload in seconds"
          example: 2.5
        previous_model:
          type: object
          nullable: true
          description: "Previous model info"
        timestamp:
          type: integer
          format: int64
          description: "Reload timestamp"
          example: 1640995200
    CurrentModelResponse:
      type: object
      required:
        - experiment_id
        - version
        - load_time
        - model_info
        - performance_stats
        - timestamp
      properties:
        experiment_id:
          type: string
          description: "Current experiment ID"
          example: "production_hmm"
        version:
          type: string
          description: "Current model version"
          example: "1.2.0"
        load_time:
          type: number
          format: float
          description: "Model load timestamp"
          example: 1640995200
        model_info:
          type: object
          description: "Model metadata"
        performance_stats:
          type: object
          description: "Model performance statistics"
        timestamp:
          type: integer
          format: int64
          description: "Response timestamp"
          example: 1640995200
    AvailableModelsResponse:
      type: object
      required:
        - models
        - total_count
        - timestamp
      properties:
        models:
          type: array
          items:
            type: object
          description: "List of available models"
        total_count:
          type: integer
          minimum: 0
          description: "Total number of available models"
          example: 5
        timestamp:
          type: integer
          format: int64
          description: "Response timestamp"
          example: 1640995200
    ErrorResponse:
      type: object
      required:
        - error
        - error_code
        - message
        - timestamp
      properties:
        error:
          type: string
          description: "Error category"
          example: "VALIDATION_ERROR"
        error_code:
          type: string
          description: "Specific error code"
          example: "INVALID_OBSERVATIONS"
        message:
          type: string
          description: "Human-readable error message"
          example: "Observations must contain exactly 3 numeric values"
        timestamp:
          type: integer
          format: int64
          description: "Error timestamp"
          example: 1640995200
        request_id:
          type: string
          description: "Request identifier if provided"
          example: "req-12345"
        details:
          type: object
          description: "Additional error details"
  responses:
    ValidationError:
      description: Validation error in request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_observations:
              summary: Invalid observation vector
              value:
                error: "VALIDATION_ERROR"
                error_code: "INVALID_OBSERVATIONS"
                message: "Observations must contain exactly 3 numeric values"
                timestamp: 1640995200
                details:
                  field: "observations"
                  provided_length: 2
                  required_length: 3
    ModelNotFound:
      description: Model not found or not loaded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            no_model_loaded:
              summary: No model currently loaded
              value:
                error: "MODEL_ERROR"
                error_code: "NO_MODEL_LOADED"
                message: "No model is currently loaded"
                timestamp: 1640995200
...
    ServiceUnavailable:
      description: Service temporarily unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            service_unavailable:
              summary: Service overloaded
              value:
                error: "SERVICE_UNAVAILABLE"
                error_code: "SERVICE_OVERLOADED"
                message: "Service is temporarily overloaded, please retry"
                timestamp: 1640995200
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (future implementation)
...